wheel 디버깅

+++ app/src/main/java/com/example/vibecapandroid/HistoryYoutubeActivity.kt
@@ -47,7 +47,7 @@ class HistoryYoutubeActivity:AppCompatActivity() {
     private val saveFolderName = "Vibecap"
     // 다운받은 파일이 저장될 위치 설정
     private val outputFilePath = Environment.getExternalStoragePublicDirectory(
-        Environment.DIRECTORY_DOWNLOADS + "/$saveFolderName"
+        Environment.DIRECTORY_DOWNLOADS + "/$saveFolderName.jpg"
     ).toString()
     private var mDownloadManager: DownloadManager? = null
     private var mDownloadQueueId: Long? = null
@@ -68,7 +68,7 @@ class HistoryYoutubeActivity:AppCompatActivity() {
         Log.d("vibe_keywords","$vibeKeyWords")
         videoID=intent.extras!!.getString("video_id")
 
-
+        var deleteCounter=0
 
         val position=intent.extras!!.getInt("position")
         Youtubeplay()
@@ -98,17 +98,24 @@ class HistoryYoutubeActivity:AppCompatActivity() {
 
         }
         viewBinding.btDelete.setOnClickListener{
-            arrayList!!.removeAt(position)
-            historyMainAdapters?.notifyItemRemoved(position)
-            deletePhoto()
+            if(deleteCounter==0) {
+                arrayList!!.removeAt(position)
+                historyMainAdapters?.notifyItemRemoved(position)
+                deletePhoto()
+                deleteCounter=1
+            }else if(deleteCounter==1){
+                Toast.makeText(
+                    applicationContext,
+                    "해당 사진이 이미 서버에서 삭제 되었습니다",
+                    Toast.LENGTH_SHORT
+                ).show()
+                deleteCounter=2
+            }
         }
 
-        viewBinding.btDownload.setOnClickListener(){
-            URLDownloading(uri);
-        }
 
     }
-    override fun onPostResume() {
+    /*override fun onPostResume() {
         super.onPostResume()
 
         // 브로드캐스트 리시버 등록
@@ -119,7 +126,7 @@ class HistoryYoutubeActivity:AppCompatActivity() {
     override fun onPause() {
         super.onPause()
         unregisterReceiver(downloadCompleteReceiver)
-    }
+    }*/
 
 
     override fun onRestart() {
@@ -201,7 +208,7 @@ class HistoryYoutubeActivity:AppCompatActivity() {
             .replace(R.id.history_youtube_you_tube_player_view, YoutubePlayerFragment)
             .commitNow()
     }
-
+/*
     //스틱코드
      open fun URLDownloading(url: Uri) {
         Log.d(TAG, "태그 URLDownloading")
@@ -223,7 +230,11 @@ class HistoryYoutubeActivity:AppCompatActivity() {
         request.setNotificationVisibility(DownloadManager.Request.VISIBILITY_VISIBLE_NOTIFY_COMPLETED) // setNotificationVisibility : VISIBILITY_VISIBLE로 설정되면 notification에 보여진다.
         request.setDestinationUri(Uri.fromFile(outputFile)) // setDestinationUri : 파일이 저장될 위치의 URI
         request.setAllowedOverMetered(true)
-
+        Toast.makeText(
+            this@HistoryYoutubeActivity,
+            "다운로드를 완료하였습니다.",
+            Toast.LENGTH_SHORT
+        ).show()
         // DownloadManager 객체 생성하여 다운로드 대기열에 URI 객체를 넣는다.
         mDownloadQueueId = mDownloadManager!!.enqueue(request)
     }
@@ -243,11 +254,8 @@ class HistoryYoutubeActivity:AppCompatActivity() {
                 val reason: Int = cursor.getInt(columnReason)
                 cursor.close()
                 when (status) {
-                    DownloadManager.STATUS_SUCCESSFUL -> Toast.makeText(
-                        this@HistoryYoutubeActivity,
-                        "다운로드를 완료하였습니다.",
-                        Toast.LENGTH_SHORT
-                    ).show()
+                    DownloadManager.STATUS_SUCCESSFUL -> {
+                        Log.d("HistoryYoutube","Download completede")}
                     DownloadManager.STATUS_PAUSED -> Toast.makeText(
                         this@HistoryYoutubeActivity,
                         "다운로드가 중단되었습니다.",
@@ -268,17 +276,17 @@ class HistoryYoutubeActivity:AppCompatActivity() {
         Log.d(TAG, "태그 갤러리 갱신 : $Image_Path")
 
         // 이전 사용 방식
-        /*Intent mediaScanIntent = new Intent(Intent.ACTION_MEDIA_SCANNER_SCAN_FILE);
+        *//*Intent mediaScanIntent = new Intent(Intent.ACTION_MEDIA_SCANNER_SCAN_FILE);
     File f = new File(Image_Path);
     Uri contentUri = Uri.fromFile(f);
     mediaScanIntent.setData(contentUri);
-    this.context.sendBroadcast(mediaScanIntent);*/
+    this.context.sendBroadcast(mediaScanIntent);*//*
         val file = File(Image_Path)
         MediaScannerConnection.scanFile(
             this, arrayOf(file.toString()),
             null, null
         )
-    }
+    }*/
 
 
 }
